+++ crates/plotinator-logs/src/mbed_motor_control/pid/entry.rs
@@ -28,7 +28,7 @@ impl fmt::Display for PidLogEntry {
 }
 
 impl PidLogEntry {
-    pub(crate) fn timestamp_ns(&self) -> f64 {
+    pub(crate) fn timestamp_ns(&self) -> i64 {
         match self {
             Self::V1(e) => e.timestamp_ns(),
             Self::V2(e) => e.timestamp_ns(),
