+++ crates/plotinator-logs/src/mbed_motor_control/status/entry/v1.rs
@@ -104,8 +104,8 @@ impl LogEntry for StatusLogEntryV1 {
         ))
     }
 
-    fn timestamp_ns(&self) -> f64 {
-        (self.timestamp_ms as u64 * 1_000_000) as f64
+    fn timestamp_ns(&self) -> i64 {
+        self.timestamp_ms as i64 * 1_000_000
     }
 }
